fix: pair review feedback
