+++ src/index.ts
@@ -3,7 +3,7 @@ export { Flyover } from './sdk/flyover'
 
 export { type FlyoverNetworks } from './constants/networks'
 
-export { type RegisterPeginParams } from './sdk/registerPegin'
+export { type RegisterPeginLbcParams } from './sdk/registerPegin'
 export { type ValidatePeginTransactionParams, type ValidatePeginTransactionOptions } from './sdk/validatePeginTransaction'
 
 export { FlyoverError } from './client/httpClient'
